requirements.txt
numpy
torch>=1.7
opencv_python
loguru
scikit-image
tqdm
torchvision>=0.10.0
Pillow
thop
ninja
tabulate
tensorboard
lap
motmetrics
filterpy
h5py
pandas
matplotlib
pycocotools
cython_bbox
dtw
seaborn
tqdm
scipy
scikit_learn
openpyxl

# verified versions
onnx==1.8.1
onnxruntime==1.8.0
onnx-simplifier==0.3.5

